Updates to use font-awesome instead of the ico-moon icons
Added dependency on font-awesome and using some of it's icons in place
of icomoon used in the inital version.  Font-awesome just seems more
widely used already and I added the ability to change out the particular
class used on the span within the template, so if the end user of the
control wants they can just override the defaults easily.